✨ Add mock implementation in gbk lib
This is controlled by a cargo feature flag `mock` and can be used to do
unit testing with the gbk lib. Currently, it is just a bare minimum
implementation that builds but all functions error. The idea is that
there will also be an interface to set what these functions should
return when running tests.