✨ Map attachment files in wasi executor
  * Wasi now supports multiple sandboxes.
  * Add two endpoints for uploading attachments, one of them streamed
through grpc and the other one is not. The reason for two endpoints is
that the setup we will use to host the server registry in gcp will not
support streamed grpcs. The local one will use streaming while the
gcp one will not.
  * Function attachments now implements a attachment download trait.
This means that attachment can be downloaded by calling attachment.download()
instead of calling code_download()