✨ Add host folder mappings
Host folder mappings is a way for the function (in its' manifest) to
express that it wants host folders to be mapped into its' filesystem
sandbox. This makes it possible to do discovery on the file system. Note
that the mounts will be read-only so it is not a giant exploit.