✨ Add get and list for functions
This implements support for listing (and filtering) and getting (by id)
functions. The functions are either stored in memory or in a Postgresql
database. All filtering functionality is implemented for both backends.

Please note that the memory backend should not be used in production
simply because it does not persist anything and do not synchronize
between instances.