✨ Add root cert store and auth to lomax
This will query the system store for certificates. Reason that we do not
bundle the app with certificates from nix is that it is bad security
practice to not stay up to date with certificates (revocations and
such).

Authentication is currently being done by reading environment
variables containing oauth tokens. This will change when we have an
oauth service later

Add middleware library for intercepting tonic requests as http.