🔨 Inputs, outputs and values are now streams and channels
A `Stream` is a collection of `Channel`s that in turn contains a list with
all the values. This means that all values are lists and when the user
asks for a single item we simply pick that item from the list if it
exists.

This also means that the specs for inputs and outputs on a function are
now unified into a `StreamSpec`. We have chosen to leave the terminology
of inputs and outputs when writing functions though. This means that you
still declare inputs in the manifest and you still call `get_input` to
get an input. This is done to avoid internal data representation choices
becoming a user-facing "feature".

This also uses built in protobuf types for everything which gets rid of
our ugly byte-fiddling.

We have also created a new rust crate with convenience methods for
working with the protobuf types, called `firm-types`. This crate also
re-exports the proto types which means that to use proto types together
with convenience methods, you only have to depend on `firm-types`.