✨ Add filesystem runtime source
A file system runtime source is represented by a directory on disk that
contains a number of runtime executables (currently only supports wasm).
File system sources also require a checksum toml file containing
checksums for all runtimes in the directory. Currently the directory is
scanned once on startup, for performance reasons. If we want to support
dynamically added runtimes a filesystem watcher can be used.

On Windows the directory `%PROGRAMDATA%/Avery/runtimes` is added if it
exists and on unix the corresponding directory is
`/usr/share/avery/runtimes`. Additional directories can be added in the
configuration file under the key `runtime_directories`.