✨ Expose all firm methods to python functions
This gives functions written in python the ability to interact with
the firm function framework. Examples could be getting inputs, mapping
attachments etc.

You can test the api when you are in the python runtime shell by
running the following command.
`runApiExample`

This will build and run two examples. The example that tests setting
errors had to be separate since it changes how the rest of the code
behaves (you can't get any function output if the function errors).

The exception is network related functionality which is not yet
implemented.