✨ Add dependency bundling for python functions
Python packages are curated in wasi-python-packages.nix so we can
guarantee that they are working for wasi. Currently the set is very
minimal.

Python packages are built into wheels and attached to the function. These
are then unpacked by the python runtime inside the wasi guest and
PYTHONPATH is adjusted accordingly.