✨ Provide avery with runtimes function
This function allows for creating a derivation that contains Avery and
the passed in runtimes. Avery is also set up to read all the runtimes
when Avery is launched.

There is also an avery with default runtimes. This is if you want to
use avery with the shipped runtimes but do not want to add any of your
own. The attribute `avery` is still the raw package.