✨ Add an internal socket for Avery
The internal socket is a unix domain socket that is only accessible
from localhost. The idea is that the TCP port is used for external
communication and the UDS is used for internal communication.

Currently only works on UNIX platforms, Windows is up next.

Also fix the deployment function:

Stop using a fixed SSL cert bundle. It is important that the SSL
bundle used gets updates even though the deploy script might
not. Therefore, we need to use the system SSL bundle.