✨ 🔨 Add generator for self signed token in Avery
The token generator will use either a provided rsa or edcsa key or
generate its own edcsa key. Changed the config for auth to be a map
of variants (none, oidc, self-signed, key-file). This maps scope (usually
the host) to an auth config.

Also the openssl dependency has been removed in favor of rustls.