✨ Add Serialization and Deserialization of tokens
This saves the serialized tokens in a file in the Avery-folder for the current user. This
results in the user only having to log in once for the providers that support this caching
mechanism (OIDC is the current example).

It also makes us less likely to hit any
rate-limits since we can use the OAuth2 refresh token instead of launching a new auth flow
every time so win-win.