✨ Add Lomax - a TCP proxy for Avery
Avery is a local-only service exposed via a trusted connection over a Unix Domain Socket
or a Windows named pipe. To allow for remote operations, Lomax runs as a system user,
accepting remote TCP connections (on port 1939 default). On an incoming connection, it
uses the information in the provided JWT to determine which local user the remote
connection is for. It then uses the Avery Auth API for that local user to determine if the
remote user has access in which case traffic is forwarded to that instance.

In cases where the connecting user does not have access, a Remote Access Request is added
through the Avery auth service. This request can be approved/denied through the Avery API
and currently Bendini is the only client that has support for interacting with the API
from the client side. This is done through the `bendini auth` set of commands.