🐛 Fix known negative caching for runtime URLs
Functions in the internal registry gets a `_dev-checksum` to use for
cache lookup since they can have the same name, version, code checksum
but still be different. Since the URL is used as the cache key for
attachment download and the path to a runtime file is the same even if
the content has change we now add an extra timestamp query parameter to
the url.