✨ Make Avery windows service compatible
- Add `--service` argument to Avery, this dispatches avery as a service
  which listens to the service control manager's STOP and START events.
- When Avery runs a service it uses the Windows Event Logger.
- Add service stop callback functionality
Due to how the service functions are set up for windows (decoupled
from any state) we do not own the function calls. This makes some
things trickier such as how to communicate certain things. In this
case we needed a trigger to be set when the service gets a
callback. The only way to make other parts of the code to find this
trigger is to make a global or send the shutdown signal method
everywhere (would have to refactor auth etc.). Decided to just have a
global for the event. At least it's only in one file and only for
windows and everything else works the same.