🐛 🏁 Fix event log message type
Windows event log expects the executable to contain information of how
to present messages from the app. This was included by mistake during
development with the winlog crate but replaced with a custom
implementation. We (temporarily hopefully) add back a dependency on
winlog and trick rust to build and link it.