✨ Add windows installer for firm applications
Currently only installs avery, there is a bunch of TODOs in the code to
be done before next firm release.

Avery is installed as a user services, the only way to start an instant
of a user service is to log in. After user has logged out and in again
we can restart the service instance during updates. The other
applications are just copied for now.