🐛 Fix windows pending deletions
Before we used the winapi to register deletions and manually edited
the registry to remove pending deletions.

Changed to manually do both operations in the registry.

Due to the nature of the registry the value of the pending file
deletions may not always be in a good state. Added somefunctionality
that tries to repair the value slightly if it is in an unexpected
state. This is because we've previously experienced that our files got
added as pending deletions in an invalid way even when using the
winapi. Now when we are doing it ourselves we need to know about the
expected format but we can at least ensure that our files get deleted
on reboot.