✨ Add hosted_domains as a config option
Support both hosted_domain and hosted_domains as a list, and if anything
in the list matches we consider it accepted. This is because the same
service can be run on a machine where users of multiple domains exists.

When multiple host domains are used the generated url will set
hd to "*".