🔒 Set windows pipe permissions
This changes named pipes in windows to have the following permissions:
RW <COMPUTER>\<username>
RW NT AUTHORITY\SYSTEM
R  <COMPUTER>\<username>-<sid>

In short, we dropped "Everyone", "Anonymous User" and "Administrators".
Lomax as a system service will be the AUTHORITY\SYSTEM user.

Unclear why we get the extra entry with sid but this is close to what we
want and might be good enough.