✨ Group function versions when listing
To get all versions for a selected function, use the `ListVersions`
endpoint instead. `List` and `ListVersions` currently use the same
filtering parameters but they are interpreted slightly different. For
`List`, the `name` filter is used for substring matching on the function
name whereas in `ListVersions` it is used to match exactly the function
name.

Filtering of metadata and publisher is done after the grouping of
versions, meaning that those filters will only be applied to the latest
version matching the version requirement when using `List`. For
`ListVersions` no grouping of versions occur so the filters will be
applied to all versions.

Also implement support for formatting options in Bendini. It now
supports three format options for `list` and `list-versions`: short,
long and json. `short` shows a condensed version of the list and `long`
provides more details on each function. `json` will output the function
list in JSON format (pretty-printed if stdout is a TTY). This can be
useful for piping to other tools. Note that this should only be used for
simple scripting purposes. For anything that does not fall into that
category, use the actual gRPC API.

Introduce `publisher_email` as a filter to both list-endpoints. This
will (as the name suggests) filter functions on the email of the person
who published the function. As stated above, this filter will, in the
case of `List` be applied after the versions have been collapsed and
will therefore only be matched against the newest version of the
function matching the version requirement.