🔨 Make auth init side-effect-free
Previously, generation of keys, identity retrieval and upload of public
keys was done on startup of Avery. Now instead, this is done as late as
possible, making initialization of the auth service free of any side
effects.

Due to how the key store is dependent on the token store for acquiring
its credentials, there is some hoop-jumping necessary to be able to have
the correct lock scopes.