🔨 Refactor login in Avery
All endpoints that would require an interactive login will now return
the gRPC status code for `unauthenticated`. This is a signal to the
client that they can try to call `login` if they have an interactive
context (think UI, CLI). `login` will generate a stream of login
commands (currently only supports Browser to signal to the client to
open a browser). This means that there are now two distinct categories
of providers: Those that can generate a new token without user
interaction and those who need user interaction. The providers that fall
in the latter category will do their work in `login` instead of
`generate`. This behavior is partly enforced since the login stream sink
is only sent to the `login` method and not to the non-interactive
counterpart `generate`.

The interactive login will currently do as many oidc logins as needed to
cover all auth scopes. When adding more "interactive providers", they
will need to do the corresponding work here.

Bendini has a new function `with_login` that can be wrapped around any
API calls to handle login if needed and retry the original request in
that case.