✨ Add wait for approval method
This makes it possible for clients to wait for approval for a given
remote access request. This is done by implementing `Future` for the
approval process. Clients can then use grpc timeouts or so to give the
approval a deadline.

Lomax now uses this functionality and if the timeout expires also
removes the pending remote access request so that users can not approve
a request that will never be used.