✨ Add support for multiple keys in keystore
Before the format in the online keystore looked like the following
  "user_name": {
    "private_key": "SomeKey"
  }

This means we only had a support for a single key which became
problematic for users using more than one computer. We changed the
format to look like the following.

  "user_name": {
    "private_keys": {
     "<key_id1>": "SomeKey",
     "<key_id2>": "AnotherKey"
     ...
    }
  }

This format ensures that we can support several keys. We added a
version file to the keys stored locally and if this version is changed
the keys are re-generated and uploaded. If we get a token with the old
key id in the header (whithout a key fingerprint) we look in the old
place in the online key store.