✨ Support hostname expansion in alt names
Support expansion of `{hostname}` to the real hostname in the
`certificate_alt_names` config setting. To properly support this, a
version of the certificate was also added that is based on a djb2 hash
of config content that impacts the generation of the certificate. If
this hash changes, the certificate is regenerated even though it exists.