🚧 ✨ Replace wasmer with wasmtime
Output redirection and compilation cache works.

Started with host API and there is one test as an example.

We now have runtimes that runs wasi or just wasm

TODO is our host API

Introduce Turbo ISLC, the compiler to generate our rust/wasm code with
wrappers and a trait to implement. This is to not have to repeat the
memory juggling and pointer sizes in wasm for every function.