⬆️ Use Nedryland 8.0.0 and Nedryglot 1.0.0
Firm is now also a nix flake, which means the version of nixpkgs can be
controlled from the user side. Nedryland's check apps are forward and
combined with Firm's binaries, for easy use with `nix run .#bendini` for
example.

Use rust from nixpkgs for stable and oxalica for nightly.
Added a rust extension to provide:
- base.languages.rust.withWasi
- base.languages.rust.nightly
- base.languages.rust.nightly.withWasi
- base.languages.rust.nativeTools (with windows cross target)

Update CI to use flakes and use workflows and actions from Nedryland.